+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project from 1.0.0 forward will be documented in thi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-01-06
+
+### Added
+
+-   Question mark supported on sql query.
+
+### Changed
+
+-   Updated to Lupdo v2.1.0.
+
+### Fixed
+
+-   Postgres Type Parser.
+
 ## [1.0.0] - 2022-12-29
 
 First Public Release On Npm